+++ ydb/tests/fq/yt/kqp_yt/cfg/kqprun_scheme.sql
