+++ ydb/tests/fq/yt/kqp_yt/cfg/test_whitelist.txt
