+++ ydb/tests/fq/yt/kqp_yt/kqp_yt_file/part0/ya.make
