+++ ydb/tests/fq/yt/kqp_yt/kqp_yt_file/part10/ya.make
