+++ ydb/tests/fq/yt/kqp_yt/kqp_yt_file/part12/ya.make
