+++ ydb/tests/fq/yt/kqp_yt/kqp_yt_file/part13/ya.make
