+++ ydb/tests/fq/yt/kqp_yt/kqp_yt_file/part17/ya.make
