+++ ydb/tests/fq/yt/kqp_yt/kqp_yt_file/part19/ya.make
