+++ ydb/tests/fq/yt/kqp_yt/kqp_yt_file/part2/ya.make
