+++ ydb/tests/fq/yt/kqp_yt/kqp_yt_file/part8/ya.make
