+++ ydb/tests/fq/yt/kqp_yt/ya.make
@@ -1,0 +1,4 @@
+RECURSE(
+    kqp_yt_file
+    kqp_yt_import
+)
