+++ ydb/tests/fq/yt/ya.make
@@ -1,4 +1,3 @@
 RECURSE(
-    kqp_yt_file
-    kqp_yt_import
+    kqp_yt
 )
